Extract GraphQL books
